+++ test/sub-module.js
@@ -1,12 +1,13 @@
 'use strict'
 
 const test = require('tape')
-const Hook = require('../')
+
+const { Hook } = require('../')
 
 test('require(\'sub-module/foo\') => sub-module/foo.js', function (t) {
   t.plan(3)
 
-  const hook = Hook(['sub-module/foo'], function (exports, name, basedir) {
+  const hook = new Hook(['sub-module/foo'], function (exports, name, basedir) {
     t.equal(name, 'sub-module/foo')
     return exports
   })
@@ -22,7 +23,7 @@ test('require(\'sub-module/foo\') => sub-module/foo.js', function (t) {
 test('require(\'sub-module/bar\') => sub-module/bar/index.js', function (t) {
   t.plan(3)
 
-  const hook = Hook(['sub-module/bar'], function (exports, name, basedir) {
+  const hook = new Hook(['sub-module/bar'], function (exports, name, basedir) {
     t.equal(name, 'sub-module/bar')
     return exports
   })
@@ -38,7 +39,7 @@ test('require(\'sub-module/bar\') => sub-module/bar/index.js', function (t) {
 test('require(\'sub-module/bar/../bar\') => sub-module/bar/index.js', function (t) {
   t.plan(3)
 
-  const hook = Hook(['sub-module/bar'], function (exports, name, basedir) {
+  const hook = new Hook(['sub-module/bar'], function (exports, name, basedir) {
     t.equal(name, 'sub-module/bar')
     return exports
   })
@@ -54,7 +55,7 @@ test('require(\'sub-module/bar/../bar\') => sub-module/bar/index.js', function (
 test('require(\'sub-module/conflict\') => sub-module/conflict.js', function (t) {
   t.plan(3)
 
-  const hook = Hook(['sub-module/conflict'], function (exports, name, basedir) {
+  const hook = new Hook(['sub-module/conflict'], function (exports, name, basedir) {
     t.equal(name, 'sub-module/conflict')
     return exports
   })
@@ -70,7 +71,7 @@ test('require(\'sub-module/conflict\') => sub-module/conflict.js', function (t) 
 test('require(\'sub-module/conflict/index.js\') => sub-module/conflict/index.js', function (t) {
   t.plan(3)
 
-  const hook = Hook(['sub-module/conflict'], function (exports, name, basedir) {
+  const hook = new Hook(['sub-module/conflict'], function (exports, name, basedir) {
     t.equal(name, 'sub-module/conflict')
     return exports
   })